backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Data Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0

# File Processing
openpyxl>=3.1.5
aiofiles>=24.1.0

# HTTP Client
httpx>=0.28.0

# Environment
python-dotenv>=1.0.1

# Monitoring
prometheus-client>=0.21.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.10.0
flake8>=7.1.0
